backend/requirements.txt
asn1crypto==0.24.0
boto==2.49.0
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
defusedxml==0.5.0
Django==2.1.2
django-braces==1.13.0
django-cors-headers==2.4.0
django-oauth-toolkit==1.2.0
django-rest-framework-social-oauth2==1.1.0
django-rest-knox==3.1.5
djangorestframework==3.8.2
future==0.16.0
idna==2.7
oauthlib==2.1.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycparser==2.19
PyJWT==1.6.4
pyOpenSSL==18.0.0
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
urllib3==1.23
